sales-insights-genai-main/requirements.txt
pandas
prophet
xgboost
scikit-learn
google-generativeai
streamlit
langchain
plotly
